project for vs2003 added (without editor) some bug with foobar fixed

+++ trunk/MyGUIEngine/include/MyGUI_LogStream.h
@@ -108,8 +108,8 @@ namespace MyGUI
 
 		inline LogStream& operator<<(size_t _value)
 		{
-			std::cout << _value;
-			if (mStream.is_open()) mStream << _value;
+			std::cout << (unsigned int)_value;
+			if (mStream.is_open()) mStream << (unsigned int)_value;
 
 			return *this;
 		}
